# docs/swagger.yaml
definitions:
  domain.Country:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
  domain.Department:
    properties:
      created_at:
        type: string
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  domain.Role:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  dto.CountryCreateDTO:
    properties:
      name:
        type: string
      status:
        type: integer
    required:
    - name
    - status
    type: object
  dto.CountryUpdateDTO:
    properties:
      name:
        type: string
      status:
        type: integer
    type: object
...
  dto.DepartmentCreateDTO:
    properties:
      location:
        type: string
      name:
        type: string
      status:
        type: integer
    required:
    - location
    - name
    - status
    type: object
  dto.DepartmentUpdateDTO:
    properties:
      location:
        type: string
      name:
        type: string
      status:
        type: integer
    required:
    - location
    - name
    - status
    type: object
...
  dto.RoleCreateDTO:
    properties:
      name:
        type: string
      status:
        type: integer
    required:
    - name
    - status
    type: object
  dto.RoleUpdateDTO:
    properties:
      name:
        type: string
      status:
        type: integer
    required:
    - name
    - status
    type: object
...
info:
  contact: {}
  description: This is a volunteer service API
  title: Onboarding and Volunteer Service API
  version: "1.0"
paths:
...
  /api/v1/countries:
    get:
      description: Get all countries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Country'
            type: array
      summary: Get all countries
      tags:
      - country
    post:
      consumes:
      - application/json
      description: Create a new country
      parameters:
      - description: Country data
        in: body
        name: country
        required: true
        schema:
          $ref: '#/definitions/dto.CountryCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Country'
      summary: Create a new country
      tags:
      - country
  /api/v1/countries/{id}:
    delete:
      description: Delete country
      parameters:
      - description: Country ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete country
      tags:
      - country
    get:
      description: Get country by ID
      parameters:
      - description: Country ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Country'
      summary: Get country by ID
      tags:
      - country
    put:
      consumes:
      - application/json
      description: Update country
      parameters:
      - description: Country ID
        in: path
        name: id
        required: true
        type: integer
      - description: Country data
        in: body
        name: country
        required: true
        schema:
          $ref: '#/definitions/dto.CountryUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Country'
      summary: Update country
      tags:
      - country
  /api/v1/departments:
    get:
      description: Get all departments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Department'
            type: array
      summary: Get all departments
      tags:
      - department
    post:
      consumes:
      - application/json
      description: Create a new department
      parameters:
      - description: Department data
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/dto.DepartmentCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Department'
      summary: Create a new department
      tags:
      - department
  /api/v1/departments/{id}:
    delete:
      description: Delete department
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete department
      tags:
      - department
    get:
      description: Get department by ID
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Department'
      summary: Get department by ID
      tags:
      - department
    put:
      consumes:
      - application/json
      description: Update department
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      - description: Department data
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/dto.DepartmentUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Department'
      summary: Update department
      tags:
      - department
  /api/v1/role/:
    get:
      description: Get all roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Role'
            type: array
      summary: Get all roles
      tags:
      - role
    post:
      description: Create role
      parameters:
      - description: Create Role Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RoleCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Role'
      summary: Create role
      tags:
      - role
  /api/v1/role/{id}:
    delete:
      description: Delete role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete role
      tags:
      - role
    get:
      description: Get role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Role'
      summary: Get role by ID
      tags:
      - role
    put:
      description: Update role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Role Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RoleUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Role'
      summary: Update role
      tags:
      - role
...
securityDefinitions:
  bearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"